feat: deep dive topics proposal (#5)
* feat: deep dive topics proposal

This is a proposal for deep dive topics, they've been sourced from various folks (you know who you are - thank you ❤️). This is a work in progress, but worth PR-ing for some feedback/review. There are a few that might need to be cut, I do not yet have an idea for an owner for _all_ of them but there's more here than we need so hopefully it should be ok!

If you have any additions to be considered, please comment here 🤗

* docs: update topic description

* docs: add topic description template

* docs: send a PR

* feat: some more awesome libp2p problems

* feat: add telemetry deep dive

* docs: add icons to package manager relevant dives

* fix: update title